requirements.txt
streamlit>=1.32.0
pyrebase4>=5.2.3
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.20.0